+++ src/context.ts
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { publicApiType } from './createApi';
+import { type publicApiType } from './createApi';
 
 export const VisibilityContext = React.createContext<publicApiType>(
   {} as publicApiType