+++ src/hooks/useItemsChanged.test.tsx
@@ -4,7 +4,7 @@ import { render } from '@testing-library/react';
 import useItemsChanged from './useItemsChanged';
 import { id as itemId } from '../constants';
 import ItemsMap from '../ItemsMap';
-import { IOItem } from '../types';
+import { type IOItem } from '../types';
 
 const TestComponent = ({
   menuItems,